Refactor how env vars are handled
This commit changes the way env variables are managed in the script.
Before all the env vars were global and we needed to load and validate them one by one on script init.
Now we introduced the `Env` class that groups them all together and has a method to load and validate them. All the vars are accessible from a single global variable.
This new class should help to make easy adding new vars.